+++ tuig/src/lib.rs
@@ -2,21 +2,6 @@
 //!
 //! If you want to jump straight in with an example, check [`docs::walkthrough`].
 //!
-//! # Architecture
-//!
-//! tuig is built around a shared event bus. Everything that happens in the game is represented by a single type you
-//! define that extends [`Message`]. You'll also have a lot of different types of [`Agent`]s, which do all of your
-//! relevant simulation and event processing. The thing the player actually interacts with is the [`Game`], which
-//! processes user input and produces the output, and communicates with agents by spawning events. Click those links
-//! for more details on each specific trait.
-//!
-//! `Agent`s and the `Game` (coincidentally the name of my new ska band) can inject new agents or messages through
-//! [`Replies`], which is the general handle into the game's internals for most things. An `Agent` can put itself to
-//! sleep -- including metaphorically -- by returning different [`ControlFlow`]s, which is useful for boosting
-//! performance by reducing how many agents need to be considered but shouldn't be relied on for anything too serious.
-//! `Game`s have a similar option in [`Response`], though they're gonna get called constantly anyway, so their
-//! responses are about re-rendering the text or exiting the game entirely.
-//!
 //! # Feature selection
 //!
 //! Before you can actually *run* your game, you need to select some backend functionality with Cargo features.
@@ -28,9 +13,8 @@
 //!     good at making good use of all available cores for this sort of thing.
 //! -   `run_single`: Useful only if `rayon` is undesirable for some reason, e.g. small-scale unit tests or `no_std`.
 //!
-//! The IO system -- which of the [`tuig-iosys::backends`](https://docs.rs/tuig-iosys/latest/tuig_iosys/backends) --
-//! will be handling our platform input and output. All the `tuig-iosys` backends are available, but the features
-//! have an extra `io_` prefix. So you can have:
+//! The IO system -- one of the [`tuig-iosys::backends`](tuig_iosys::backends) -- will be handling our platform input
+//! and output. All the `tuig-iosys` backends are available, but the features have an extra `io_` prefix. So you have:
 //!
 //! -   `io_nop`: Great for integration tests where you don't actually care about input or output, but want an
 //!     otherwise complete tuig.
@@ -39,8 +23,28 @@
 //!     It's very widely compatible, because you literally don't need any 3D hardware for it to work.
 //!
 //! You have to pick exactly one runner, but you can choose more than one IO system. [`Runner::load_run`] will try to
-//! intelligently pick "the best it can" given the ones you've turned on, but if you very reasonably disagree, you can
-//! load your preferred system and use [`Runner::run`] instead.
+//! intelligently pick "the best it can" given the ones you've turned on, but if you very reasonably disagree, or if
+//! you want a third-party backend, you can load your preferred system and use [`Runner::run`] instead.
+//!
+//! # Architecture
+//!
+//! tuig is built around a shared event bus. Everything that happens in the game is represented by a single type 
+//! <code>M: [Message]</code>. You'll also have a variety of of [`Agent<M>`]s, which do all of your actual simulation
+//! and event processing. The thing the player actually interacts with is the [`Game<M>`], which processes user input
+//! and renders the output, and communicates with agents by spawning events. Click those links for more details on
+//! each specific trait.
+//! 
+//! Events passed around are loosely organized into "rounds". As events are passed to agents, their [`Replies`] are
+//! collected, then applied only after each agent has seen the current round of events. However, agents **aren't
+//! necessarily** running in lockstep. They don't even all necessarily see the same rounds! You'll occasionally see
+//! them mentioned because that's how the engine works internally, and it explains why certain things happen or don't.
+//!
+//! `Agent`s and the `Game` (coincidentally the name of my new ska band) can inject new agents or messages through
+//! [`Replies`], which is the general handle into the game's internals for most things. An `Agent` can put itself to
+//! sleep -- including metaphorically -- by returning different [`ControlFlow`]s, which is useful for boosting
+//! performance by reducing how many agents need to be considered but shouldn't be relied on for anything too serious.
+//! `Game`s have a similar option in [`Response`], though they're gonna get called constantly anyway, so their
+//! responses are about re-rendering the text or exiting the game entirely.
 
 #![cfg_attr(doc, feature(doc_cfg, doc_auto_cfg))]
 tuig_pm::force_docs_nightly!();
@@ -51,10 +55,12 @@ pub mod docs;
 mod game;
 mod runner;
 mod util;
+mod message;
 
 pub use {
     agent::{Agent, ControlFlow, WaitHandle},
-    game::{Game, Message, Replies, Response},
+    game::{Game, Response},
+    message::{Message, Replies},
     runner::Runner,
     tuig_iosys as io,
 };
