+++ tuig/src/util/mod.rs
@@ -1,2 +1,4 @@
+//! Miscellaneous internal-only utilities.
+
 pub mod macros;
 pub mod timing;
